refactor: fix pointers as params and helpers folder
