+++ src/application/controller/user/create_user.go
@@ -2,12 +2,12 @@ package userController
 
 import (
 	"github.com/gofiber/fiber/v2"
-	"github.com/rodrigoRVSN/beeus-api/src/domain/entity"
-	fieldsValidator "github.com/rodrigoRVSN/beeus-api/src/infra/helpers/validator"
+	"github.com/rodrigoRVSN/beeus-api/src/application/dto"
+	fieldsValidator "github.com/rodrigoRVSN/beeus-api/src/infra/helpers"
 )
 
 func (controller *UserController) CreateUser(context *fiber.Ctx) error {
-	payload := new(entity.User)
+	payload := new(dto.SignUpInputDTO)
 
 	if err := context.BodyParser(&payload); err != nil {
 		return context.Status(fiber.StatusBadRequest).JSON(fiber.Map{"status": "fail", "message": err.Error()})
@@ -19,7 +19,7 @@ func (controller *UserController) CreateUser(context *fiber.Ctx) error {
 		return context.Status(fiber.StatusBadRequest).JSON(errors)
 	}
 
-	err := controller.useCase.CreateUser(payload)
+	err := controller.useCase.CreateUser(*payload)
 
 	if err != nil {
 		return context.Status(fiber.StatusBadRequest).JSON(fiber.Map{"status": "fail", "message": err.Error()})
