+++ src/application/use_case/user/create_user.go
@@ -1,15 +1,17 @@
 package userUseCase
 
-import "github.com/rodrigoRVSN/beeus-api/src/domain/entity"
+import (
+	"github.com/rodrigoRVSN/beeus-api/src/application/dto"
+)
 
-func (uc *CreateUserUseCase) CreateUser(user *entity.User) error {
-	userWasFound := uc.UserGateway.GetUserByEmail(user, user.Email)
+func (uc *CreateUserUseCase) CreateUser(payload dto.SignUpInputDTO) error {
+	userWasFound := uc.UserGateway.CheckIfUserAlreadyExists(payload, payload.Email)
 
 	if userWasFound != nil {
 		return userWasFound
 	}
 
-	err := uc.UserGateway.CreateUser(user)
+	err := uc.UserGateway.CreateUser(payload)
 
 	if err != nil {
 		return err
