+++ src/infra/helpers/hash.go
@@ -1,4 +1,4 @@
-package hash
+package helpers
 
 import "golang.org/x/crypto/bcrypt"
